refactor(App): control strip to its own component
Split up the app component to facilitate more override points.

ToolStrip has been renamed to ControlsStrip to avoid name conflicts with
the tools.
