+++ src/components/tools/ResetViews.vue
@@ -1,6 +1,6 @@
 <script lang="ts">
 import { defineComponent } from 'vue';
-import ToolButton from '@/src/components/ToolButton.vue';
+import ControlButton from '@/src/components/ControlButton.vue';
 import { createEventHook } from '@vueuse/core';
 
 const clickEvent = createEventHook<void>();
@@ -10,7 +10,7 @@ export function useResetViewsEvents() {
 
 export default defineComponent({
   components: {
-    ToolButton,
+    ControlButton,
   },
   // setup() {
   // },
@@ -23,7 +23,7 @@ export default defineComponent({
 </script>
 
 <template>
-  <tool-button
+  <control-button
     size="40"
     icon="mdi-camera-flip-outline"
     name="Reset Views"
